(#15) Changed interface for Domain timings

+++ test/Schur/PatchIfaceScatter_MPI2.cpp
@@ -176,8 +176,8 @@ TEST_CASE(
 		}
 	}
 
-	CHECK(local_vector->getNumLocalPatches() == patch_iface_interfaces.size());
-	CHECK(local_vector->getNumLocalCells() == n * patch_iface_interfaces.size());
+	CHECK(local_vector->getNumLocalPatches() == (int) patch_iface_interfaces.size());
+	CHECK(local_vector->getNumLocalCells() == n * (int) patch_iface_interfaces.size());
 }
 TEST_CASE(
 "Schur::PatchIfaceScatter<2> getNewLocalPatchIfaceVector returns vector with local MPI_Comm",
